myTasks To-Do list project, first Push !
This is the first puch of my 2-days challenge, myTasks, which i decided to continue developping after my 48 hours to code an app challenge, and, even if it is not released yet, it's going on and it will be out soon, when all the features will be ready. The app icon is also missing... Anyway i hope this app will be good, and will be different from all the other To-Do list apps out on Google Play. Stay tuned, and try myTasks !